requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings ==2.7.1
aiofiles ==24.1.0
langchain== 0.3.14
PyMuPDF== 1.25.1
langchain-community== 0.3.14
motor==3.6.1
openai== 1.61.0
cohere== 5.13.11
qdrant-client==1.13.2
milvus==2.3.5
weaviate-client== 4.11.0
orjson== 3.10.15
scikit-learn